requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib
bcrypt
pydantic
pydantic[email]
python-jose
python-multipart
pdfplumber
groq